improved error handling on config file

+++ src/env.js
@@ -132,12 +132,17 @@ let config = {
     }
 };
 
-try {
-    config = yaml.safeLoad(fs.readFileSync(vector.configFile, 'utf8')) || config;
-} catch(e) {
+
+if (fs.existsSync(vector.configFile)) {
+    try {
+        config = yaml.safeLoad(fs.readFileSync(vector.configFile, 'utf8')) || config;
+    } catch (error) {
+        throw new Error("The file " + vector.configFile + " is not valid yml: " + error.message.split(":")[0]);
+    }
+
+} else {
     console.log("Impossible to load config.yml. A default configuration file has been generated.");
     fs.writeFileSync(defaultConfigFilePath, yaml.dump(config))
-
 }
 
 const formatLine = printf(({ level, message, label, timestamp }) => `${timestamp} [${label}] ${level}: ${message}`);
